changing hashing and security approach
now, GET endpoint is fully public - if someone got a hash - they have access. HMAC to prevent brute-force attacks

+++ src/routes/send-ai-request.ts
@@ -1,93 +1,76 @@
-import { FastifyPluginAsync, FastifySchema } from 'fastify';
-import { Static, Type } from '@sinclair/typebox';
+import { FastifyPluginAsync } from 'fastify';
 import { getPromptHash } from '../helpers/hashing.js';
-import { checkCacheEntryExists, checkCacheErrorExists, createCacheEntry, deleteCacheError, getAssistant, getCurrentUser } from '../services/directus.js';
+import { checkCacheEntryExists, createCacheEntry, deleteCacheError, getAssistant, getCurrentUser } from '../services/directus.js';
 import { getJob, shelveJob } from '../services/jobs.js';
 import { getCompletion } from '../services/ai.js';
 import { v4 as uuidv4 } from 'uuid';
+import { FileInfo } from '../helpers/file-info.js';
 
-const QueryString = Type.Object({
-  q: Type.String({ minLength: 1 }),
-  a: Type.String({ minLength: 1 }),
-  t: Type.String({ minLength: 10 }),
-})
+interface FieldInfo {
+   type: 'field' | 'file'
+   mimetype: string
+}
 
-type QueryStringType = Static<typeof QueryString>
+interface FileFieldInfo extends FieldInfo {
+  toBuffer: () => Promise<Buffer>
+}
 
-const schema: FastifySchema = {
-  querystring: QueryString
+interface TextFieldInfo extends FieldInfo {
+  value: string
+}
+
+// Fastify schema does not work well with file uploads, so downgrading to pure TS, compile-time-only typing
+interface RequestBody {
+  type: 'field' | 'file'
+  value?: string
+  toBuffer?: () => Promise<Buffer>
+}
+
+function isFile(request: FieldInfo): request is FileFieldInfo {
+  return request.type === 'file';
+}
+
+function isField(request: FieldInfo): request is TextFieldInfo {
+  return request.type === 'field';
+}
+
+interface Body {
+  a: FieldInfo
+  t: FieldInfo
+  q: FieldInfo
+}
+
+const getString = (field: FieldInfo) => {
+  if (!isField(field)) {
+    throw new Error('Field is not a string');
+  }
+  if (!field.value) {
+    throw new Error('Field is missing value');
+  }
+
+  return field.value
 };
 
-interface IHeaders {
-  
-}
+const getFile = async (field: FieldInfo): Promise<string | FileInfo> => {
+  if (isField(field)) {
+    return getString(field);
+  }
+  if (!isFile(field)) {
+    throw new Error('Invalid field type');
+  }
 
-interface IReply {
-  
-}
+  const buffer = await field.toBuffer();
+  return {
+    contentBase64: buffer.toString('base64'),
+    mimeType: field.mimetype,
+  };
+};
 
 const plugin: FastifyPluginAsync = async (fastify, opts) => {
-  fastify.get<{
-    Querystring: QueryStringType,
-    Headers: IHeaders,
-    Reply: IReply
-  }>('/', { schema } , async (request, reply) => {
-    const { q, a, t } = request.query
-
-    const { id: userId } = await getCurrentUser(t);
-    const {
-      id: assistantId,
-      sys_message,
-      name: assistant_name,
-      ai_config,
-      date_updated: assistantUpdatedAt,
-      format,
-   } = await getAssistant(t, a);
-    const promptHash = getPromptHash(userId, assistantUpdatedAt, q);
-
-    if (!await checkCacheEntryExists(t, promptHash) && !getJob(promptHash)) {
-      shelveJob(promptHash, (async () => {
-        let content: string;
-        try {
-          content = await getCompletion(sys_message, q, ai_config);
-          await deleteCacheError(t, promptHash);
-          await createCacheEntry(t, {
-            hash: promptHash,
-            prompt: q.trim(),
-            content,
-            status: 'content',
-            title: `${assistant_name} | ${q.trim().substring(0, 20)}`,
-            format,
-          });
-        } catch (er: any) {
-          request.log.error(er);
-          await deleteCacheError(t, promptHash);
-          await createCacheEntry(t, {
-            hash: promptHash,
-            prompt: q.trim(),
-            content: '<mark>Could not get completion from assistant.</mark>',
-            status: 'error',
-            title: `${assistant_name} | ${q.trim().substring(0, 20)}`,
-            format,
-          });
-        }
-      })());
-    }
-
-    const expires = new Date();
-    expires.setDate(expires.getDate() + 365);
-    reply.setCookie('t', t, {
-      expires,
-     });
-    reply.redirect(`/${promptHash}`);
-  });
-
-  fastify.post<{ Headers: {}; Reply: {} }>('/', {  }, async (request, reply) => {
-    const data = await request.file();
-    if (!data) return reply.status(400).send({ error: 'Missing form fields' });
-
-    const { a: { value: a }, t: { value: t }, q: file } = data.fields as any;
-    const q = new Blob([await file.toBuffer()], { type: data.mimetype });
+  fastify.post<{ Body: Body }>('/', {  }, async (request, reply) => {
+    const t = getString(request.body.t);
+    const a = getString(request.body.a);
+    const q = await getFile(request.body.q);
 
     const { id: userId } = await getCurrentUser(t);
     const {
@@ -99,42 +82,40 @@ const plugin: FastifyPluginAsync = async (fastify, opts) => {
       format,
     } = await getAssistant(t, a);
 
-    const stub = uuidv4();
-    const promptHash = getPromptHash(userId, assistantUpdatedAt, stub);
+    const prompt = typeof q === 'string' ? q : '(File Input)';
+    const title = `${assistant_name} | ${prompt.trim().substring(0, 20)}`;
+    const promptHash = await getPromptHash(userId, assistantUpdatedAt, typeof q === 'string' ? q : q.contentBase64);
 
-    shelveJob(promptHash, (async () => {
-      let content: string;
-      try {
-        content = await getCompletion(sys_message, q, ai_config, format === 'file' ? 'file' : undefined);
-        await deleteCacheError(t, promptHash);
-        await createCacheEntry(t, {
-          hash: promptHash,
-          prompt: stub,
-          content,
-          status: 'content',
-          title: `${assistant_name} | File Request`,
-          format,
-        });
-      } catch (er: any) {
-        request.log.error(er);
-        await deleteCacheError(t, promptHash);
-        await createCacheEntry(t, {
-          hash: promptHash,
-          prompt: stub,
-          content: '<mark>Could not get completion from assistant.</mark>',
-          status: 'error',
-          title: `${assistant_name} | File Request`,
-          format
-        });
-      }
-    })());
+    if (!await checkCacheEntryExists(promptHash) && !getJob(promptHash)) {
+      shelveJob(promptHash, (async () => {
+        let content: string;
+        try {
+          content = await getCompletion(sys_message, q, ai_config, format === 'file' ? 'file' : undefined);
+          await deleteCacheError(t, promptHash);
+          await createCacheEntry(t, {
+            hash: promptHash,
+            prompt,
+            content,
+            status: 'content',
+            title,
+            format,
+          });
+        } catch (er: any) {
+          request.log.error(er);
+          await deleteCacheError(t, promptHash);
+          await createCacheEntry(t, {
+            hash: promptHash,
+            prompt,
+            content: '<mark>Could not get completion from assistant.</mark>',
+            status: 'error',
+            title,
+            format: 'markdown',
+          });
+        }
+      })());
+    }
 
-    const expires = new Date();
-    expires.setDate(expires.getDate() + 365);
-    reply.setCookie('t', t, {
-      expires,
-    });
-    reply.redirect(`/${promptHash}`);
+    reply.code(201).send({ id: promptHash });
   });
 }
